requirements.txt
alembic==1.12.1
annotated-types==0.6.0
anyio==3.7.1
boto3==1.29.6
botocore==1.32.6
certifi==2023.11.17
charset-normalizer==3.3.2
exceptiongroup==1.2.0
fastapi==0.104.1
idna==3.5
importlib-metadata==6.8.0
importlib-resources==6.1.1
jmespath==1.0.1
Mako==1.3.0
MarkupSafe==2.1.3
mysql==0.0.3
mysql-connector-python==8.2.0
mysqlclient==2.2.0
protobuf==4.21.12
pydantic==2.5.2
pydantic-settings==2.1.0
pydantic_core==2.14.5
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
s3transfer==0.7.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
typing_extensions==4.8.0
urllib3==1.26.7
zipp==3.17.0
